Split utils into fileutils and ci
